Move ProcessManager.ProcessedUntil to processinfo.go
